construct extrinsic and extrinsic proof for grandpa header
